+++ flytekit/common/tasks/task.py
@@ -1,4 +1,3 @@
-import copy as _copy
 import hashlib as _hashlib
 import json as _json
 import logging as _logging
@@ -63,7 +62,6 @@ class SdkTask(
             custom,
             container=container,
         )
-        self._has_fast_registered = False
 
     @property
     def interface(self):
@@ -169,32 +167,6 @@ class SdkTask(
         except Exception:
             self._id = old_id
             raise
-
-    @_exception_scopes.system_entry_point
-    def fast_register(self, project, domain, name, digest, additional_distribution) -> str:
-        """
-        :param Text project: The project in which to register this task.
-        :param Text domain: The domain in which to register this task.
-        :param Text name: The name to give this task.
-        :param Text digest: The version in which to register this task.
-        :param Text additional_distribution: User-specified location for remote source code distribution.
-        :rtype: Text: Registered identifier.
-        """
-
-        original_container = self.container
-        container = _copy.deepcopy(original_container)
-        args = ["pyflyte-fast-execute", "--additional-distribution", additional_distribution, "--"] + container.args
-        container._args = args
-        self._container = container
-
-        try:
-            registered_id = self.register(project, domain, name, digest)
-        except Exception:
-            self._container = original_container
-            raise
-        self._has_fast_registered = True
-        self._container = original_container
-        return str(registered_id)
 
     @property
     def has_fast_registered(self) -> bool:
